Address review on #4862: move attemptOption to NonEmptyAlternative + add law
  - Move `attemptOption` from `Alternative` to `NonEmptyAlternative`
    (satorg #1).  Implementation only needs `combineK + map + pure`.
  - Use `widen(fempty)` with a cached `F[Option[Nothing]]` to avoid
    allocating a fresh `pure(None)` on every call (johnynek nit).
  - Add `nonEmptyAlternativeAttemptOptionConsistentWithCombineKAndPure`
    to `NonEmptyAlternativeLaws` and wire it into the `nonEmptyAlternative`
    rule set (satorg #3).  This required adding `EqFOA: Eq[F[Option[A]]]`
    to the `nonEmptyAlternative` and `alternative` rule sets' implicit
    parameter lists.
  - Drop the redundant point test in `AlternativeSuite` (satorg inline).
  - Move the two `attemptOption` property tests to `NonEmptyAlternativeSuite`
    to follow the method's new location.
  - Add two `mima.sbt` exclusions for the test-helper signature change.
